// frontend/src/app/modules/smart-action/enums/smart-action-actions.enum.ts
export enum SmartActionActionNames {
  SUBMIT_COMMAND = '[Smart Action] Submit Command',
  PARSE_SUCCESS = '[Smart Action] Parse Success',
  PARSE_FAILURE = '[Smart Action] Parse Failure',
  EXECUTE = '[Smart Action] Execute',
  EXECUTE_SUCCESS = '[Smart Action] Execute Success',
  EXECUTE_FAILURE = '[Smart Action] Execute Failure',
  UNDO = '[Smart Action] Undo',
  CLEAR_RESULT = '[Smart Action] Clear Result',
}